Server/requirements.txt
# Core dependencies
asgiref>=3.6.0
click==8.1.3
Django>=4.1.9
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
django-cors-headers>=4.3.0
requests==2.31.0

# Database
pymongo==4.6.1
dnspython==2.3.0
sqlparse>=0.4.4

# Authentication and Security
bcrypt==4.1.2
PyJWT==2.8.0

# Natural Language Processing
nltk==3.7
regex>=2023.12.25
joblib==1.2.0
deep-translator==1.11.4
httpx==0.13.3
httpcore==0.9.1

# Utilities
tqdm==4.64.1
setuptools==67.8.0

# Additional dependencies
gunicorn==21.2.0
Werkzeug==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5

# Development dependencies
pytest==7.4.0
pytest-django==4.5.2
black==23.7.0
flake8==6.1.0